backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
websockets==14.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.1
alembic==1.14.0
aiosqlite==0.20.0
greenlet==3.1.1
